+++ tests/FSharp.Data.Tests/JsonParserProperties.fs
@@ -1,8 +1,8 @@
 ﻿#if INTERACTIVE
 #r "../../bin/lib/net45/FSharp.Data.dll"
-#r "../../packages/NUnit/lib/net45/nunit.framework.dll"
+#r "../../packages/test/NUnit/lib/net45/nunit.framework.dll"
 #r "../../packages/FsCheck/lib/net45/FsCheck.dll"
-#r "../../packages/FsUnit/lib/net46/FsUnit.NUnit.dll"
+#r "../../packages/test/FsUnit/lib/net46/FsUnit.NUnit.dll"
 #else
 module FSharp.Data.Tests.JsonParserProperties
 #endif