+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 #### 2.4.6 - March 25 2018
 * Added `ContentTypeWithEncoding` helper to `HttpRequestHeaders`.
 * `JsonValue` will explicitly set content type charset to UTF-8 when making requests.
+* Prevent superfluous encoding of URL parameters.
 
 #### 2.4.5 - February 19 2018
 * Add an optional parameter for the `System.Text.Encoding` to use when reading data to the CSV, HTML, and Json providers. This parameter is called `encoding` and should be present on all Load and AsyncLoad methods.