+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+#### 3.0.0-beta4 - Unreleased
+* (Breaking Change) Ignore culture when parsing JSON to better match the JSON spec.
+
 #### 3.0.0-beta3 - April 9 2018
 * Increased type caches TTL from 10 seconds to 5 minutes.